Still trying to get to an agreement between Sargent2005's data (EPSC: duration 2.5ms, amplitude 50pA, charge transfer 120fC) with Jason's (g/eLeak, V_threshold, V_reset...).
At the moment, the cell is not reactive enough, but I've just noticed that my charge tranfer value is around 35fC. Actually, my amplitude agrees with Sargent, while my (average) time course is not too far from Jason's. This is probably not a good thing because the shapes of the waveforms are very different. Now that I've acknowledged this limitation in my model, the next step will be to tweak the time course to bring it closer to the value given in Sargent2005. If that doesn't work, I may try tweaking again the amplitude as well. I'd keep changing the waveform shape as a last resort.