Plot Jason's actual waveform fits. Add FG analysis script.
There must be some problem with the port of Jason's model that I'm
using to plot as a comparison after the optimisation procedure, so I
just asked him to explicitly calculate the traces for me. For the time
being, only the AMPA traces are there, but NMDA should arrive soon.